Added some shtuff..
I added a main page (index.html)... I know the css needs a little bit more work.
I have not linked anything to eachother, but will be my main goal tomorrow. 

I also managed to make the cards for the actual maps and a button to click on them. 
I am having a hard time getting the background image working on the trail.html

Any suggestions you have will be great. For index.html, I messed with the font for the privacy and the only way I could get it to show up a little bit more noticable was to change the opacity of the background image.

YALL WE DOIN THIS!!!!!!!!